[UHYU-259] type: 유저 타입 중복 정의 제거 및 일원화
