+++ src/features/mypage/api/types.ts
@@ -1,30 +1,16 @@
 import type { ApiResponse } from '@/shared/client/client.type';
+import type { UserInfoData } from '@/shared/types';
 
-export type UserRole = 'USER' | 'ADMIN';
-export type UserStatus = 'ACTIVE' | 'DELETED';
-export type UserGrade = 'GOOD' | 'VIP' | 'VVIP';
-export type Gender = 'MALE' | 'FEMALE';
-
-export interface Marker {
-  id: number;
-  createdAt: string;
-  updatedAt: string;
-  markerImage: string;
-}
-
-// Postman API 응답과 완전히 일치하는 데이터 구조
-export interface UserInfoData {
-  profileImage: string;
-  userName: string;
-  nickName: string | null;
-  email: string;
-  age: number | null;
-  gender: Gender | null;
-  grade: UserGrade | null;
-  brandIdList?: number[];
-  updatedAt: string;
-  role?: UserRole; // 백엔드 구현 전, USER 데이터 추가 필요
-}
+// 공통 타입들은 @/shared/types에서 import하여 사용
+export type {
+  UserRole,
+  UserGrade,
+  UserGender,
+  Marker,
+  UserInfo,
+  UserInfoData,
+  UpdateUserRequest,
+} from '@/shared/types';
 
 // 개인정보 수정 API 응답 데이터 구조
 export interface UpdateUserResponseData {
@@ -34,33 +20,6 @@ export interface UpdateUserResponseData {
 // API 응답 타입들 - 공통 ApiResponse 타입 사용
 export type UserInfoResponse = ApiResponse<UserInfoData>;
 export type UpdateUserResponse = ApiResponse<UpdateUserResponseData>;
-
-// 기존 UserInfo 타입 (내부 사용용)
-export interface UserInfo {
-  id: number;
-  createdAt: string;
-  updatedAt: string;
-  userName: string;
-  nickName: string;
-  kakaoId: number;
-  email: string;
-  age: number;
-  gender: Gender;
-  role: UserRole;
-  status: UserStatus;
-  grade: UserGrade;
-  profileImage: string;
-  age_range?: string;
-  marker: Marker;
-  brandIds: number[];
-}
-
-// 개인정보 수정 요청 타입 (실제 수정 가능한 필드들)
-export interface UpdateUserRequest {
-  updatedNickName?: string;
-  updatedGrade?: UserGrade;
-  updatedBrandIdList?: number[];
-}
 
 // 즐겨찾기 관련 타입
 export interface Bookmark {
