+++ src/features/user/api/types.ts
@@ -1,29 +1,11 @@
-export type UserGrade = 'GOOD' | 'VIP' | 'VVIP';
+// 공통 타입들은 @/shared/types에서 import하여 사용
+export type {
+  UserGrade,
+  UserGender,
+  UserRole,
+  UserExtraInfoRequest,
+  CheckEmailRequest,
+} from '@/shared/types';
 
-export type UserGender = 'MALE' | 'FEMALE';
-export type UserRole = 'USER' | 'ADMIN';
-export type UserStatus = 'ACTIVE' | 'INACTIVE' | 'SUSPENDED';
-
-export interface UserInfomation {
-  profileImage: string;
-  userName: string;
-  nickName: string | null;
-  email: string;
-  age: number | null;
-  gender: UserGender | null;
-  grade: UserGrade | null;
-  markerId: number | null;
-  role?: UserRole;
-  updatedAt: string;
-}
-
-export interface UserExtraInfoRequest {
-  grade: UserGrade;
-  recentBrands: number[];
-  interestedBrands: number[];
-}
-
-// 이메일 중복 확인 관련 타입
-export interface CheckEmailRequest {
-  email: string;
-}
+// UserInfomation은 UserInfoData로 통합 (기존 호환성 유지를 위해 alias 제공)
+export type { UserInfoData as UserInfomation } from '@/shared/types';
