+++ src/features/user/index.ts
@@ -2,3 +2,16 @@ export * from './api/userApi';
 export * from './api/types';
 export * from './api/endpoints';
 export * from './hooks/useUserMutation';
+
+// 공통 타입들을 re-export (기존 호환성 유지)
+export type {
+  UserRole,
+  UserGrade,
+  UserGender,
+  UserInfo,
+  UserInfoData,
+  SimpleUserInfo,
+  UpdateUserRequest,
+  UserExtraInfoRequest,
+  CheckEmailRequest,
+} from '@/shared/types';
